deduplicated TestAnalyzer and TestAttribute
